+++ test.sh
@@ -15,7 +15,9 @@ javac --release 8 -Xlint:-options -d "$OUT" \
   "$ROOT/src/com/dumbphone/powertoggle/PowerPrecondition.java" \
   "$ROOT/tests/com/dumbphone/powertoggle/ToggleStateTest.java" \
   "$ROOT/tests/com/dumbphone/powertoggle/ToggleCommandTest.java" \
-  "$ROOT/tests/com/dumbphone/powertoggle/PowerPreconditionTest.java"
+  "$ROOT/tests/com/dumbphone/powertoggle/PowerPreconditionTest.java" \
+  "$ROOT/tests/com/dumbphone/powertoggle/AirplaneModePlanTest.java"
 java -cp "$OUT" com.dumbphone.powertoggle.ToggleStateTest
 java -cp "$OUT" com.dumbphone.powertoggle.ToggleCommandTest
 java -cp "$OUT" com.dumbphone.powertoggle.PowerPreconditionTest
+java -cp "$OUT" com.dumbphone.powertoggle.AirplaneModePlanTest